requirements.txt
hexs>=1.1.0
python-bitcoinlib
pyparsing
txsc
